introduce a new polkit-default-privs easy profile
Since our standard profile has become too annoying for typical
single-user desktop systems in some instances (e.g. for Flatpak), we are
introducing a new easy profile that targets exactly this type of users.

For the moment profile easy == standard, but as new actions will be
whitelisted and new complaints are coming in we will adjust the easy
profile accordingly.