add whitelisting file for files installed in polkit-1/rules.d
For the new whitelist we're using JSON as a config file format. This way
we can easily import the data into the Python based rpmlint-check. A
downside of JSON is that we can't use free style comments in it.
Therefore we're using explicit comment entries in the dictionaries. This
way we can also explicitly keep track of the audit bug numbers.