restrictive profile: allow local users to reboot/shutdown (bsc#1171115)
There are many other ways to reboot or shutdown a system e.g. via the
display manager or the text login console, power button etc. To really
restrict this (e.g. for kiosk systems) a system integrator would need to
apply a lot of configuration adjustment to prevent local users to
perform these actions.

Also these settings are now in line with "login1.hibernate" and
"login1.suspend" which are already allowed for local sessions.