libvirt: allow the active session to create read-only connections (bsc#1179126)
via polkit rules members of the group libvirt are allowed to create
read-write connections to libvirt without authentication. read-only
conenctions are still restricted to root, however. This is inconsistent.

By allowing the active session to access the read-only socket without
password also in the restrictive profile this should be mitigated.